requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.4.0
blinker==1.9.0
blis==1.3.3
cachetools==6.2.2
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
cloudpathlib==0.23.0
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.13
fonttools==4.60.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
murmurhash==1.0.15
narwhals==2.12.0
nltk==3.9.2
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.5.0
preshed==3.0.12
protobuf==6.33.1
pyarrow==21.0.0
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
rpds-py==0.27.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
streamlit==1.50.0
tenacity==9.1.2
thinc==8.3.9
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1
zipp==3.23.0
